added ICP semantics

+++ src/Registration/gps.h
@@ -9,7 +9,7 @@ public:
     GPS();
     virtual ~GPS();
 
-    void createMap(std::string currentPath,Eigen::Matrix4f imu2base,Eigen::Matrix4f pc2base,Eigen::Matrix4f gps2base, IO::Datacontainer imuContainer, IO::pCloudcontainer pcContainer, IO::Datacontainer gpsUTMContainer, std::string filename, float leafSize);
+    void createMap(std::string currentPath,Eigen::Matrix4d imu2base,Eigen::Matrix4d pc2base,Eigen::Matrix4d gps2base, IO::Datacontainer imuContainer, IO::pCloudcontainer pcContainer, IO::Datacontainer gpsUTMContainer, std::string filename, float leafSize);
 
     //bool compareFunction (std::string a, std::string b);
